requirements.txt
aws==0.2.5
azure-cognitiveservices-speech==1.24.2
bcrypt==4.0.1
boto==2.49.0
boto3==1.26.13
botocore==1.29.13
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.3
docopt==0.6.2
fabric==2.7.1
google-api-core==2.10.2
google-auth==2.14.1
google-cloud==0.34.0
google-cloud-core==2.3.2
google-cloud-speech==2.16.2
google-cloud-storage==2.6.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
grpcio==1.50.0
grpcio-status==1.50.0
idna==3.4
invoke==1.7.3
jiwer==2.5.1
jmespath==1.0.1
Levenshtein==0.20.2
num2words==0.5.12
paramiko==2.12.0
pathlib2==2.3.7.post1
prettytable==3.5.0
proto-plus==1.22.1
protobuf==4.21.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyNaCl==1.5.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pywatchman==1.4.1
rapidfuzz==2.13.2
requests==2.28.1
rev-ai==2.17.1
rsa==4.9
s3transfer==0.6.0
six==1.16.0
termcolor==2.1.0
Unidecode==1.3.6
urllib3==1.26.12
wcwidth==0.2.5
websocket-client==0.59.0
